% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{components}
\alias{components}
\title{Streetmix components}
\format{
A length 4 nested list with specifications for lanes, markings,
objects, and vehicles.
}
\usage{
components
}
\description{
Streetmix components
}
\keyword{datasets}
