% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_street.R
\name{get_street}
\alias{get_street}
\title{Get data on a Streetmix street}
\usage{
get_street(
  url = NULL,
  user_id = NULL,
  street_id = NULL,
  count = NULL,
  return = "data"
)
}
\arguments{
\item{url}{Streetmix url to get data on}

\item{user_id}{Streetmix user account ID}

\item{street_id}{Streetmix street ID}

\item{count}{Count of streets to return}

\item{return}{Type of data to return based on url: "data", "street",
"location", or "segments". If \code{NULL}, return full response.}
}
\description{
Get data on a Streetmix street
}
