% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{illustrations}
\alias{illustrations}
\title{StreetMix Illustrations}
\format{
A data frame with 257 rows and 6 variables:
\describe{
\item{\code{repo}}{repository}
\item{\code{name}}{illustration name}
\item{\code{url}}{SVG file url}
\item{\code{category}}{category/folder}
\item{\code{side}}{facing side (left or right)}
\item{\code{direction}}{direction of movement (inbound or outbound)}
}
}
\usage{
illustrations
}
\description{
Data frame with links to SVG files with StreetMix illustrations \url{https://github.com/streetmix/illustrations}
}
\details{
Illustrations available under an attribution-share alike license \url{https://github.com/streetmix/illustrations/blob/main/LICENSE}

For design guidelines and colors see the illustration guide \url{https://docs.streetmix.net/contributing/illustrations/design-principles}
}
\keyword{datasets}
