% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people}
\alias{people}
\title{Streetmix CC-BY-SA illustrations of people}
\format{
A data frame with 36 rows and 5 variables:
\describe{
  \item{\code{id}}{image id}
  \item{\code{width}}{image width}
  \item{\code{name}}{image name}
  \item{\code{disallowFirst}}{logical}
  \item{\code{weight}}{integer}
}
}
\usage{
people
}
\description{
Streetmix CC-BY-SA illustrations of people
}
\keyword{datasets}
