% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{components}
\alias{components}
\title{Streetmix components}
\format{
A length 5 nested list with specifications for lanes, markings,
objects, vehicles, and effects.
}
\source{
\url{https://github.com/streetmix/streetmix/raw/refs/heads/main/client/src/segments/components.json}
}
\usage{
components
}
\description{
Last updated: 2024-11-10
}
\keyword{datasets}
