% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{components}
\alias{components}
\title{Streetmix components}
\format{
A length 5 nested list with specifications for lanes, markings,
objects, and vehicles.
}
\source{
\url{https://raw.githubusercontent.com/streetmix/streetmix/main/assets/scripts/segments/components.json}
}
\usage{
components
}
\description{
Last updated: 2024-01-08
}
\keyword{datasets}
